requirements.txt
flask
firebase-admin
google-cloud-storage
reportlab
requests
gunicorn
